+++ src/hardware.rs
@@ -16,6 +16,7 @@ pub use joypad::Joypad;
 pub use timer::Timer;
 
 use crate::{
+  flags::is_flag_set,
   hardware::{
     apu::{Apu, AudioSample},
     cartridge::{Cartridge, Mbc1, RomOnly},
@@ -280,9 +281,27 @@ impl Hardware {
     (self.interrupts.enabled_bitfield() & self.interrupts.requested_bitfield()) != 0
   }
 
-  /// Checks if the following interrupt has been requested.
-  pub fn is_interrupt_requested(&self, interrupt: Interrupt) -> bool {
-    self.interrupts.is_enabled(interrupt) && self.interrupts.is_requested(interrupt)
+  /// Returns the next pending interrupt to be handled, if any.
+  pub fn next_pending_interrupt(&self) -> Option<Interrupt> {
+    let pending = self.interrupts.enabled_bitfield() & self.interrupts.requested_bitfield();
+    // Make sure that the upper 3 bits are not set
+    let pending = pending & 0b0001_1111;
+
+    // Interrupts with lower bit values have higher priority
+    if is_flag_set!(pending, Interrupt::VBlank as u8) {
+      Some(Interrupt::VBlank)
+    } else if is_flag_set!(pending, Interrupt::Lcd as u8) {
+      Some(Interrupt::Lcd)
+    } else if is_flag_set!(pending, Interrupt::Timer as u8) {
+      Some(Interrupt::Timer)
+    } else if is_flag_set!(pending, Interrupt::Serial as u8) {
+      Some(Interrupt::Serial)
+    } else if is_flag_set!(pending, Interrupt::Joypad as u8) {
+      Some(Interrupt::Joypad)
+    } else {
+      // This case can only happen if the pending interrupts is 0.
+      None
+    }
   }
 
   /// Clears a requested [`Interrupt`].